% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alpha_.R
\name{add_alpha_}
\alias{add_alpha_}
\title{Add transparency to a color value}
\usage{
add_alpha_(col, alpha = 1)
}
\arguments{
\item{col}{Input color string.}

\item{alpha}{Transparency level, a numeric from 0 to 1 inclusive.}
}
\description{
Add transparency to a color value
}
