% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{as.exstra_score}
\alias{as.exstra_score}
\title{Convert a compatable object to the exstra_score class}
\usage{
as.exstra_score(x, copy = FALSE)
}
\arguments{
\item{x}{An object with a class that inherits from exstra_score}

\item{copy}{Should the object be copied? Slower if TRUE, but in-place data.table
operations will not change both objects. This option is here to remind 
users that this is normally copied by reference.}
}
\description{
Convert a compatable object to the exstra_score class
}
