% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{copy.exstra_score}
\alias{copy.exstra_score}
\title{Copy an exstra_score object so referencing does not cause problems.}
\usage{
\method{copy}{exstra_score}(x)
}
\description{
Copy an exstra_score object so referencing does not cause problems.
}
