% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_tsum.R
\name{copy.exstra_tsum}
\alias{copy.exstra_tsum}
\title{Copy an exstra_tsum object}
\usage{
\method{copy}{exstra_tsum}(x)
}
\arguments{
\item{x}{exstra_tsum object to copy.}
}
\description{
Prevents changing both objects on changes by reference, that do not copy on write.
}
