% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{copy}
\alias{copy}
\title{Copy an object with a data.table.}
\usage{
copy(x)
}
\arguments{
\item{x}{Object to copy.}
}
\description{
Copy-on-write doesn't apply to data.table. 
Instead, an explicit copy() must be used.
}
