% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{dim.exstra_score}
\alias{dim.exstra_score}
\title{Dimension of exstra_score object}
\usage{
\method{dim}{exstra_score}(x)
}
\arguments{
\item{x}{An exstra_score object.}
}
\value{
Vector of length two, the number of loci and number of samples respectively.
}
\description{
Dimension of exstra_score object
}
