% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exstra_known.R
\docType{data}
\name{exstra_known}
\alias{exstra_known}
\title{Known repeat expansion disorder loci.}
\format{An exstra_db object:
\describe{
  \item{db}{A data.table of the known loci}
  \item{input_type}{Indication that the loci are referred to by name (as opposed to genomic position)}
}}
\source{
Rick M. Tankard, Martin B. Delatycki, Paul J. Lockhart, 
        Melanie Bahlo. 
        Detecting known repeat expansions with standard protocol next generation 
        sequencing, towards developing a single screening test for neurological repeat 
        expansion disorders. 
        \emph{bioRxiv} 157792; 
        doi: \url{https://doi.org/10.1101/157792}
}
\usage{
exstra_known
}
\description{
Table of known repeat expansion disorder loci.
}
\keyword{datasets}
