% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_tsum.R
\name{exstra_tsum_new_}
\alias{exstra_tsum_new_}
\title{Create a new exstra_tsum object.}
\usage{
exstra_tsum_new_(strscore, T, p.values = NULL, qmats = NULL, xecs = NULL,
  args = NULL)
}
\description{
Create a new exstra_tsum object.
}
