% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_low_scores.R
\name{filter_low_scores}
\alias{filter_low_scores}
\title{Filter read scores with lower than expected scores}
\usage{
filter_low_scores(strscore)
}
\arguments{
\item{strscore}{exstra_score object.}
}
\description{
Under the assumption each base in the sequence is uniform and independent.
This may help to filter reads not overlapping the STR.
}
