% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sex.R
\name{filter_sex}
\alias{filter_sex}
\title{filter rep_score_data by sex}
\usage{
filter_sex(strscore, sex = "known", safe = TRUE)
}
\arguments{
\item{strscore}{exstra_score object}

\item{sex}{the sex to filter, can be any of:
- "all":     no filtering
- "male":    only male samples
- "female":  only female samples
- "missing": only missing samples
- "known":   only samples with sex assigned}

\item{safe}{When TRUE, missing sex assignments with cause an error when \code{sex} is any of "all", "male" or "female".}
}
\description{
filter rep_score_data by sex
}
