% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.exstra_score.R
\name{ggplot.exstra_score}
\alias{ggplot.exstra_score}
\title{Default ggplot method for exstra_score}
\usage{
\method{ggplot}{exstra_score}(data = NULL, ...)
}
\arguments{
\item{data}{exstra_score object}

\item{...}{Additional arguments to ggplot()}
}
\description{
Default ggplot method for exstra_score
}
\examples{
library(ggplot2)
data(exstra_wgs_pcr_2)
ggplot(exstra_wgs_pcr_2, aes(x = rep, colour = sample)) +
  stat_ecdf() +
  facet_wrap("locus")

}
