% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_db.R
\name{is.exstra_db}
\alias{is.exstra_db}
\title{check if the object is part of the exstra_db class}
\usage{
is.exstra_db(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
check if the object is part of the exstra_db class
}
