% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{is.exstra_score}
\alias{is.exstra_score}
\title{Test if object is an exstra_score object}
\usage{
is.exstra_score(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
Logical
}
\description{
Test if object is an exstra_score object
}
