% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_db.R
\name{loci_min_exp}
\alias{loci_min_exp}
\title{Give the minimum expanded size of the STR}
\usage{
loci_min_exp(x, locus)
}
\arguments{
\item{x}{Object that inherits from exstra_db}

\item{locus}{Locus name to get the reference size from.}
}
\description{
This function may not work as intended may be deleted.
}
