% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_db.R
\name{loci_text_info.exstra_db}
\alias{loci_text_info.exstra_db}
\title{Give text info for the locus}
\usage{
\method{loci_text_info}{exstra_db}(x, locus)
}
\arguments{
\item{x}{object of class exstra_db}

\item{locus}{Locus to get the information of}
}
\description{
Usually used in plot titles.
}
