% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{loci_text_info}
\alias{loci_text_info}
\title{Give text info for the locus}
\usage{
loci_text_info(x, locus)
}
\arguments{
\item{x}{object of class exstra_db}

\item{locus}{Locus to get the information of}
}
\description{
Usually used in plot titles.
}
