% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsum_test.R
\name{p_value_sd_}
\alias{p_value_sd_}
\title{P-value standard deviation}
\usage{
p_value_sd_(p, B, N)
}
\arguments{
\item{p}{vector of p-values}

\item{B}{Number of simulations}

\item{N}{Number of tested samples in each simulation}
}
\description{
Accounts for dependencies within each simulation
}
