% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{plot.exstra_score}
\alias{plot.exstra_score}
\title{Plot ECDF curves from an exstra_score object}
\usage{
\method{plot}{exstra_score}(x, loci = NULL, sample_col = NULL,
  refline = TRUE, ylab = "Fn(x)", verticals = TRUE, pch = 19, xlim,
  ylim = c(0, 1), alpha_control = 0.5, alpha_case = NULL,
  xlinked = "all", xlinked.safe = TRUE, x_upper_missing = 150,
  main = NULL, ...)
}
\arguments{
\item{x}{exstra_score object.}

\item{loci}{character; Only plot at the loci specified/ 
No effect if NULL.}

\item{sample_col}{Specify the colours for the given samples as a named character vector.
names() are sample names, while the values are characters specifying colours.
If specified, other samples are black by default, but may have an alpha value
less than 1 for transparency.}

\item{refline}{logical; if TRUE, then vertical lines are drawn indicating the reference size. 
This does not take into account mismatches in the STR, or false repeat hits.}

\item{ylab}{Label for the y-axis.}

\item{verticals}{logical; if TRUE, vertical lines are drawn at steps. See \code{\link{plot.stepfun}}.}

\item{pch}{numeric; the plotting character size.}

\item{xlim}{numeric of length 2; the x-limits for the plot.}

\item{ylim}{numeric of length 2; the y-limits for the plot.}

\item{alpha_control}{Transparency alpha value for the control samples.}

\item{alpha_case}{Transparency alpha value for the case samples. No effect if NULL.}

\item{xlinked}{If not "all", limits X chromosome loci to only "male" or "female" samples.
Can also filter on only samples with "known" or "missing" sex.}

\item{xlinked.safe}{logical; if TRUE when xlinked is filtering on "male" or "female", 
an error will be raised if any samples have an undefined sex. 
If FALSE, samples with missing sex will also be filtered when xlinked is "male" or "female".}

\item{x_upper_missing}{If xlim is not specified, then loci with no data will have this
upper x-axis limit.}

\item{...}{Further arguments to the \code{plot.ecdf} function.}
}
\description{
Plot ECDF curves from an exstra_score object
}
\examples{
plot(exstra_wgs_pcr_2["HD"])

plot(exstra_wgs_pcr_2, "HD", sample_col = c("WGSrpt_10" = "red", "WGSrpt_12" = "blue"))

# X-linked disorder, show male samples only:
plot(exstra_wgs_pcr_2, "SBMA", xlinked = "male")

}
