% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{plot_names}
\alias{plot_names}
\title{Give the plot names for given sample names}
\usage{
plot_names(x, names)
}
\arguments{
\item{x}{Object}

\item{names}{Sample names to get the plot names for.}
}
\description{
Give the plot names for given sample names
}
