% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsum_test.R
\name{qm_tsum_stat_bare_}
\alias{qm_tsum_stat_bare_}
\title{simple tsum statistic by known}
\usage{
qm_tsum_stat_bare_(qm, bmu, bstde)
}
\arguments{
\item{qm}{Quantile matrix (must be a matrix)}

\item{bmu}{Background mu vector}

\item{bvar}{Background variance vector}
}
\description{
Assume input has been checked, distribution trimming, and minimum quant chopped.
Background mu and variance is given for each quantile.
}
