% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_score_list.R
\name{rbind_score_list}
\alias{rbind_score_list}
\alias{rbind_score}
\title{Combine multiple exSTRa data objects.}
\usage{
rbind_score_list(strscore_list, idcol = "data_group",
  allow_sample_clash = FALSE, fill = FALSE)

rbind_score(..., idcol = "data_group", allow_sample_clash = FALSE,
  fill = FALSE)
}
\arguments{
\item{strscore_list}{A list containing \code{exstra_score} objects.}

\item{idcol}{The name of the column giving the names of \code{strscore_list} in the output.}

\item{allow_sample_clash}{If TRUE, allows a sample name to occur multiple times in the 
\code{exstra_score} objects, otherwise duplicates cause an error.}

\item{fill}{If TRUE, missing columns are filled with NAs. Default FALSE.}

\item{...}{\code{exstra_score} objects to combine.}
}
\value{
An exstra_score object.
}
\description{
Allows data from multiple \code{exstra_score} objects to be combined, similarly to
\code{\link[data.table]{rbindlist}} (data.table) and \code{\link[base]{rbind}}.
The \code{exstra_score} objects may either be given directly to \code{rbind_score}, or
as a list to \code{rbind_score_list} that is easier to use when the number of objects
to combine is not known in advance.
}
\examples{
# create a list of distinct samples
score_list <- list(group1 = exstra_wgs_pcr_2[, "WGSrpt_20"], 
                   controls = exstra_wgs_pcr_2[, c("WGSrpt_02", "WGSrpt_04")],
                   group2 = exstra_wgs_pcr_2[, c("WGSrpt_15", "WGSrpt_16")]
                  )
combined_scores <- rbind_score_list(score_list)
combined_scores$samples

# Without using a list:
rbind_score(group1 = exstra_wgs_pcr_2[, "WGSrpt_20"], 
            controls = exstra_wgs_pcr_2[, c("WGSrpt_02", "WGSrpt_04")],
            group2 = exstra_wgs_pcr_2[, c("WGSrpt_15", "WGSrpt_16")]
            )$samples

# Combining with a repeated sample name, possibly due to multiple experiments.
# Splitting data for this example:
score_1 <- exstra_wgs_pcr_2["DM1", "WGSrpt_15"]
score_2 <- exstra_wgs_pcr_2["HD", "WGSrpt_15"]
# Combine with the sample name repeated across exstra_score objects:
rbind_score_list(list(score_1, score_2), allow_sample_clash = TRUE)
# or
rbind_score(score_1, score_2, allow_sample_clash = TRUE)
}
\seealso{
\code{\link[data.table]{rbindlist}}, \code{\link[base]{rbind}}
}
