% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exstra_db.R
\name{read_exstra_db}
\alias{read_exstra_db}
\title{Read a file of known STR loci for analysis}
\usage{
read_exstra_db(file, ...)
}
\arguments{
\item{file}{Path of the file to be read.}

\item{...}{Extra arguments to the functions \code{read.delim} (text) or \code{xlsx::read.xlsx} (for xlsx files).}
}
\value{
An exstra_db object.
}
\description{
Reads a file formatted for exSTRa that contains information of short-tandem repeat
loci that may be expanded.
}
\details{
Files are assumed to be tab-delimited unless the extension is xlsx in which case the 
xlsx package is used (if installed).
}
\examples{
read_exstra_db(system.file("extdata", "repeat_expansion_disorders.txt", package = "exSTRa")) 

}
\seealso{
\code{\link{read_score}}
}
