% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exstra_db_txt_guesser.R
\name{read_exstra_db_txt_guesser}
\alias{read_exstra_db_txt_guesser}
\title{Guess the type of data inside a txt file by its header}
\usage{
read_exstra_db_txt_guesser(file, ...)
}
\arguments{
\item{file}{File path to check.}
}
\description{
Guess the type of data inside a txt file by its header
}
