% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exstra_db_ucsc.R
\name{read_exstra_db_ucsc}
\alias{read_exstra_db_ucsc}
\title{Read simple repeats track that's been downloaded from UCSC Genome Browser.}
\usage{
read_exstra_db_ucsc(file, header, ...)
}
\arguments{
\item{file}{File path.}

\item{header}{Explicitly indicate is there is a header in the file. Determined automatically otherwise.}

\item{...}{Additional arguments to read.delim()}
}
\description{
Read simple repeats track that's been downloaded from UCSC Genome Browser.
}
