% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_score.R
\name{read_score}
\alias{read_score}
\title{Read repeat score data produced from Bio::STR::exSTRa}
\usage{
read_score(file, database, groups.regex = NULL, groups.samples = NULL,
  filter.low.counts = TRUE)
}
\arguments{
\item{file}{Path of file to read.}

\item{database}{Path of database file (character) or exstra database (exstra_db object).}

\item{groups.regex}{Sets the sample groups with a regular expression. 
This is a named character vector of R regular expressions,
where names correspond to the group and regular expressions to 
match the sample names. 
The first match in the vector takes priority over other matches, 
such that it may be sensible to have the final regex as the empty
string: "". 
Group names should only be 'case', 'control' or 'null'.}

\item{groups.samples}{Sets the sample groups with a named list of character vectors, with names the group
and vectors of sample names. 
Group names should only be 'case', 'control' or 'null'.}

\item{filter.low.counts}{If TRUE, apply automatic filtering of counts lower than
expected by chance (assuming independent and uniform DNA base distribution).}
}
\description{
\code{read_score} reads output from the Perl module Bio::STR::exSTRa.
}
\examples{
str_score <- read_score (
    file = system.file("extdata", "HiSeqXTen_WGS_PCR_2.txt", package = "exSTRa"), 
    database = system.file("extdata", "repeat_expansion_disorders.txt", package = "exSTRa"), 
    groups.regex = c(control = "^WGSrpt_0[24]$", case = ""), 
    filter.low.counts = TRUE
)
str_score
str_score$samples

# Plot the HD locus only:
plot(str_score["HD"])


# Defining cases by sample name directly:
str_score_HD_cases <- read_score (
    file = system.file("extdata", "HiSeqXTen_WGS_PCR_2.txt", package = "exSTRa"), 
    database = system.file("extdata", "repeat_expansion_disorders.txt", package = "exSTRa"),
    groups.samples = list(case = c("WGSrpt_10", "WGSrpt_12")),
    filter.low.counts = TRUE
)
str_score_HD_cases
str_score_HD_cases$samples

plot(str_score_HD_cases["HD"])

# for greater control, use object from read_exstra_db() instead
str_db <- read_exstra_db(
            system.file("extdata", "repeat_expansion_disorders.txt", package = "exSTRa")
          )
str_score <- read_score (
    file = system.file("extdata", "HiSeqXTen_WGS_PCR_2.txt", package = "exSTRa"), 
    database = str_db, 
    groups.regex = c(control = "^WGSrpt_0[24]$", case = ""), 
    filter.low.counts = TRUE
)
str_score
str_score$samples

}
\references{
Bio::STR:exSTRa Perl module
\url{https://github.com/bahlolab/Bio-STR-exSTRa}
}
