% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bam.R
\name{score_bam}
\alias{score_bam}
\title{Give read scores of BAM files}
\usage{
score_bam(
  paths,
  database,
  sample_names = NULL,
  groups.regex = NULL,
  groups.samples = NULL,
  filter.low.counts = TRUE,
  scan_bam_flag = scanBamFlag(isUnmappedQuery = FALSE, isSecondaryAlignment = FALSE,
    isNotPassingQualityControls = FALSE, isDuplicate = FALSE),
  qname = FALSE,
  method = c("overlap", "count"),
  verbosity = 1
)
}
\arguments{
\item{paths}{Paths to BAM files}

\item{database}{Path of database file (character) or exstra database (exstra_db object).}

\item{sample_names}{Sample names in the same order as paths.}

\item{groups.regex}{Sets the sample groups with a regular expression. 
This is a named character vector of R regular expressions,
where names correspond to the group and regular expressions to 
match the sample names. 
The first match in the vector takes priority over other matches, 
such that it may be sensible to have the final regex as the empty
string: "". 
Group names should only be 'case', 'control' or 'null'.}

\item{groups.samples}{Sets the sample groups with a named list of character vectors, with names the group
and vectors of sample names. 
Group names should only be 'case', 'control' or 'null'.}

\item{filter.low.counts}{If TRUE, apply automatic filtering of counts lower than
expected by chance (assuming independent and uniform DNA base distribution).}

\item{scan_bam_flag}{Sets read filters based on SAM flags. If not NULL, an object returned by Rsamtools::scanBamFlag().}

\item{qname}{If TRUE, the query name of reads is given in output.}

\item{verbosity}{Control amount of messages, an interger up to 2.}
}
\description{
Requires the RSamtools package from Bioconductor to be installed. 
(May not work on M1 Macs)
}
