% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bam.R
\name{score_count_method}
\alias{score_count_method}
\title{Score DNA sequences using the count method}
\usage{
score_count_method(seqs, motif)
}
\arguments{
\item{seqs}{A vector of DNA sequences}

\item{motif}{A single DNA repeat motif string}
}
\description{
The count method counts each occurrence of the motif and its cycles
(starting on different bases).
}
