% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{[.exstra_score}
\alias{[.exstra_score}
\title{Extract loci or samples frome exstra_score object}
\usage{
\method{[}{exstra_score}(x, loc, samp)
}
\arguments{
\item{x}{exstra_score object}

\item{loc}{Select loci, using data.table filtering on x$db.}

\item{samp}{Select samples, using data.table filtering on x$samples.}
}
\value{
exstra_score object
}
\description{
Using \code{i} (select) syntax of data.table to extract loci and/or samples.
The first index is the loci filter on x$db, and second sample filter on x$samples.
}
\examples{
data(exstra_wgs_pcr_2)
# All data:
exstra_wgs_pcr_2

# Extract one locus:
exstra_wgs_pcr_2["HD"]

# Extract one sample:
exstra_wgs_pcr_2[, "WGSrpt_10"]

# Extract all triplet repeats
exstra_wgs_pcr_2[unit_length == 3]
exstra_wgs_pcr_2[unit_length == 3]$db$locus

# Extract all coding repeats
exstra_wgs_pcr_2[gene_region == "coding"]
exstra_wgs_pcr_2[gene_region == "coding"]$db$locus

# Extract all case samples:
exstra_wgs_pcr_2[, group == "case"]

# Extract by index 
exstra_wgs_pcr_2[2, 5]

}
