% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_tsum.R
\name{[.exstra_tsum}
\alias{[.exstra_tsum}
\title{Extract loci or samples from exstra_tsum object}
\usage{
\method{[}{exstra_tsum}(x, loc, samp)
}
\arguments{
\item{x}{exstra_score object}

\item{loc}{Select loci, using data.table filtering on x$db.}

\item{samp}{Select samples, using data.table filtering on x$samples.}
}
\value{
exstra_tsum object
}
\description{
Using \code{i} (select) syntax of data.table to extract loci and/or samples.
The first index is the loci filter on x$db, and second sample filter on x$samples.
}
\examples{
# Run tsum_test()
(tsum <- tsum_test(exstra_wgs_pcr_2[c("HD", "SCA2", "SCA6", "FRDA")], B = 100))

# All data:
tsum

# Extract one locus:
tsum["HD"]

# Extract one sample:
tsum[, "WGSrpt_10"]

# One sample at one locus:
tsum["HD", "WGSrpt_10"]

# Extract by index 
tsum[2, 5]

# The following are intended to work, but do not at present:
# Extract all triplet repeats
## tsum[unit_length == 3]
## tsum[unit_length == 3]$db$locus

# Extract all coding repeats
## tsum[gene_region == "coding"]
## tsum[gene_region == "coding"]$db$locus

# Extract all case samples:
## tsum[, group == "case"]

}
