% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested_exstra_pipeline.R
\name{suggested_exstra_pipeline}
\alias{suggested_exstra_pipeline}
\title{Create a graph of the suggested pipeline}
\usage{
suggested_exstra_pipeline(file = NULL, show.with.file = FALSE)
}
\arguments{
\item{file}{Output SVG file name, if required.}

\item{show.with.file}{If output to a file, determines whether to display in R also. Has no effect when file is NULL.}
}
\value{
The loci of \code{x}.
}
\description{
Create a graph of the suggested pipeline
}
\examples{
suggested_exstra_pipeline()
}
\seealso{
\pkg{\link{DiagrammeR}}

\pkg{\link{DiagrammeRsvg}}
}
