% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alpha.R
\name{add_alpha}
\alias{add_alpha}
\title{Add transparency to a color value}
\usage{
add_alpha(col, alpha = 1)
}
\arguments{
\item{col}{Input color character.}

\item{alpha}{Transparency level, a numeric from 0 to 1 inclusive.}
}
\value{
A character vector.
}
\description{
Add transparency to a color value
}
