% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_score.R
\name{copy.exstra_score}
\alias{copy.exstra_score}
\title{Copy an exstra_score object.}
\usage{
\method{copy}{exstra_score}(x)
}
\arguments{
\item{x}{exstra_score object to copy.}
}
\description{
Prevents changing both objects on changes by reference, that do not copy on write.
}
