% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exstra_known.R
\docType{data}
\name{exstra_known}
\alias{exstra_known}
\title{Known repeat expansion disorder loci.}
\format{An exstra_db object:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.biorxiv.org/content/early/2017/06/30/157792}
}
\usage{
exstra_known
}
\description{
Table of known repeat expansion disorder loci.
}
\keyword{datasets}
