% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.exstra_score.R
\name{ggplot.exstra_score}
\alias{ggplot.exstra_score}
\title{Default ggplot method for exstra_score}
\usage{
\method{ggplot}{exstra_score}(data = NULL, ...)
}
\description{
Default ggplot method for exstra_score
}
