% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_tsum.R
\name{is.exstra_tsum}
\alias{is.exstra_tsum}
\title{Check if an object is an instance of the exstra_tsum class.}
\usage{
is.exstra_tsum(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
Logical
}
\description{
Checks for the class attribute only. 
Does not check for correctness.
}
