% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{loci}
\alias{loci}
\title{Return the loci of the selected object}
\usage{
loci(x, ...)
}
\arguments{
\item{x}{object of an exSTRa class.}
}
\value{
The loci of \code{ex}.
}
\description{
Return the loci of the selected object
}
\examples{
loci(exstra_known)
loci(exstra_wgs_pcr_2)
}
