% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{Return a data.table with p-values of a tsum_exstra object.}
\usage{
p_values(tsum, correction = c("bf", "locus", "uncorrected"), alpha = 0.05,
  only.signif = FALSE, modify = FALSE, p.matrix = NULL)
}
\arguments{
\item{tsum}{An exstra_tsum object.}

\item{correction}{Correction method to use. Use "bf" or TRUE for Bonferroni correction, and 
"uncorrected" or FALSE for no correction. ("bonferroni" is also acceptable).
"locus" is Bonferroni correction by locus.}

\item{alpha}{Significance level alpha.}

\item{modify}{If TRUE, will modify the tsum$stats table. Effectively ignored if only.sig == TRUE.}

\item{p.matrix}{Matrix of p-values for internal use. Should only be used without tsum.}

\item{only.sig}{If TRUE, only return significant results.}
}
\value{
A data.table
}
\description{
Return a data.table with p-values of a tsum_exstra object.
}
