% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{Return a data.table with p-values of a tsum_exstra object.}
\usage{
p_values(tsum, correction = c("bf", "locus", "uncorrected"),
  alpha = 0.05, only.signif = FALSE, modify = FALSE,
  p.matrix = NULL)
}
\arguments{
\item{tsum}{An exstra_tsum object.}

\item{correction}{Correction method to use. Use "bf" or TRUE for Bonferroni correction, and 
"uncorrected" or FALSE for no correction. ("bonferroni" is also acceptable).
"samples" is Bonferroni correction by the number of tests (samples) at each locus.
"loci" is Bonferroni correction by the number of loci.}

\item{alpha}{Significance level alpha.}

\item{modify}{If TRUE, will modify the tsum$stats table. Effectively ignored if only.sig == TRUE.}

\item{p.matrix}{Matrix of p-values for internal use. Should only be used without tsum.}

\item{only.sig}{If TRUE, only return significant results.}
}
\value{
A \code{data.table} keyed by "locus" then "sample". 
        Other columns are \code{tsum} as calculated by \code{\link{tsum_test}}, \code{p.value} (uncorrected),
        \code{signif} (TRUE if significant after given correction), and 
        \code{p.value.sd}, giving the standard deviation of the p-value estimate from the 
        simulation.
}
\description{
Return a data.table with p-values of a tsum_exstra object.
}
