% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{Return a data.table with p-values of a tsum_exstra object.}
\usage{
p_values(tsum, correction = c("bf", "uncorrected"), alpha = 0.05,
  only.signif = FALSE)
}
\arguments{
\item{tsum}{An exstra_tsum object}

\item{correction}{Correction method to use. Use "bf" for Bonferroni correction, and 
"uncorrected" for no correction. ("bonferroni" is also acceptable)}

\item{alpha}{Significance level alpha.}

\item{only.sig}{If TRUE, only return significant results.}
}
\value{
A data.table
}
\description{
Return a data.table with p-values of a tsum_exstra object.
}
