% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_tsum.R
\name{plot.exstra_tsum}
\alias{plot.exstra_tsum}
\title{Plot exstra_tsum, highlighting significant results}
\usage{
\method{plot}{exstra_tsum}(
  x,
  loci = NULL,
  sample_col = NULL,
  correction = NULL,
  alpha = NULL,
  controls_label = "Not significant",
  alpha_nonsignif = 0.25,
  ...
)
}
\arguments{
\item{x}{An exstra_tsum object.}

\item{loci}{Character vector of locus or loci to plot.}

\item{sample_col}{Named (samples) vector of charcters defining colors.}

\item{correction}{Apply this correction method from \code{\link{p_values}}. 
If NULL, then the correction already applied to tsum is used.}

\item{alpha}{Significance level. 
If NULL, then the significance already applied to tsum is used. (default 0.05)}

\item{controls_label}{Generic label for all control samples.}

\item{alpha_nonsignif}{Transparency alpha value for nonsignificant samples.}

\item{...}{further arguments to \code{\link{plot.exstra_score}}}
}
\description{
Plot the significant results of a tsum
}
\seealso{
\code{\link{plot.exstra_score}}
}
