% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_exstra_tsum.R
\name{plot.exstra_tsum}
\alias{plot.exstra_tsum}
\title{Plot exstra_tsum, highlighting significant results}
\usage{
\method{plot}{exstra_tsum}(tsum, loci = NULL, sample_col = NULL,
  correction = "bf", alpha = 0.05, controls_label = "Not significant",
  alpha_nonsignif = 0.25, ...)
}
\arguments{
\item{tsum}{An exstra_tsum object.}

\item{loci}{Character vector of locus or loci to plot.}

\item{sample_col}{Named (samples) vector of charcters defining colors.}

\item{...}{further arguments to plot.exstra_score}
}
\description{
Plot the significant results of a tsum
}
\seealso{
plot.exstra_score
}
