% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.R
\name{plot_multi}
\alias{plot_multi}
\title{Multiple STR score ECDF plots}
\usage{
plot_multi(strscore, prefix = "exSTRa_plot", dir = "images/",
  plot_cols = NULL, color_only = NULL, plottypes = 1,
  alpha_control = 0.5, alpha_case = NULL, legend = TRUE,
  legend_control = TRUE, controls_label = "controls",
  custom_legend = NULL, ...)
}
\description{
Allows plotting the STR scores of many loci, including direct to files.
}
