% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_bam.R
\name{score_bam}
\alias{score_bam}
\title{Give read scores of BAM files}
\usage{
score_bam(
  paths,
  database,
  sample_names = NULL,
  sample_name_origin = c("RG", "basename"),
  sample_name_remove = "",
  sample_name_extract = ".+",
  groups.regex = NULL,
  groups.samples = NULL,
  filter.low.counts = TRUE,
  scan_bam_flag = Rsamtools::scanBamFlag(isUnmappedQuery = FALSE, isSecondaryAlignment =
    FALSE, isNotPassingQualityControls = FALSE, isDuplicate = FALSE),
  qname = FALSE,
  method = c("overlap", "count"),
  parallel = FALSE,
  cluster_n = NULL,
  cluster = NULL,
  verbosity = 1
)
}
\arguments{
\item{paths}{Paths to BAM files}

\item{database}{Path of database file (character) or exstra database (exstra_db object).}

\item{sample_names}{Sample names in the same order as paths. May be set to NULL, 
in which case sample names are found bound sample_name_origin.}

\item{sample_name_origin}{Option to determine sample names from, either the BAM read group (RG) tag, or path basename. Ignored if sample_names is not NULL.}

\item{sample_name_remove}{If sample_names == NULL, regular expression passed to stringr::str_remove to 
remove parts of sample names. 
".bam" extensions are automatically removed.}

\item{sample_name_extract}{If sample_names == NULL, regular expression passed to stringr::str_extract to only use part of sample names.}

\item{groups.regex}{Sets the sample groups with a regular expression. 
This is a named character vector of R regular expressions,
where names correspond to the group and regular expressions to 
match the sample names. 
The first match in the vector takes priority over other matches, 
such that it may be sensible to have the final regex as the empty
string: "". 
Group names should only be 'case', 'control' or 'null'.}

\item{groups.samples}{Sets the sample groups with a named list of character vectors, with names the group
and vectors of sample names. 
Group names should only be 'case', 'control' or 'null'.}

\item{filter.low.counts}{If TRUE, apply automatic filtering of counts lower than
expected by chance (assuming independent and uniform DNA base distribution).}

\item{scan_bam_flag}{Sets read filters based on SAM flags. If not NULL, an object returned by Rsamtools::scanBamFlag().}

\item{qname}{If TRUE, the query name of reads is given in output.}

\item{method}{Read score method. 
"overlap" tags bases that form part of a repeat motif (starting on any base), then counts the bases that are tagged with any starting base of the motif. 
"count" simply counts the number of matches of the motif, cycled on its starting bases. 
The original method of the original exSTRa publication closely resembles "overlap", 
though some differences from Bio::STR::exSTRa is expected to occur. 
The "count" method is simplier and therefore probably faster, and may possibly be better. 
We are still evaluating which method is superior.}

\item{parallel}{Use the parallel package when simulating the distribution, creating the
required cluster. 
If cluster is specified then this option makes no difference.}

\item{cluster_n}{If parallel is TRUE, then the number of nodes in the cluster is 
 automatically set as half of those available on your machine 
 (but never less than 1). This option allows manual setting of the 
 number of nodes, either less to free up other resources, or to 
 maximize available resources. 
If cluster is specified then this option makes no difference.}

\item{cluster}{A snow cluster created by snow::makeCluster(). Use if you wish to set up 
the cluster yourself or reuse an existing cluster.}

\item{verbosity}{Control amount of messages, an interger up to 2.}
}
\description{
Requires the RSamtools package from Bioconductor to be installed. 
(May not work on M1 Macs)
}
