% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsum_test_speedy.R
\name{tsum_statistic_1locus}
\alias{tsum_statistic_1locus}
\title{Determine tsum_statistic_1locus assuming input data is for only one locus}
\usage{
tsum_statistic_1locus(strscore_loc, case_control = FALSE,
  min.quant = 0, trim = 0, give.pvalue = FALSE, parallel = FALSE,
  cluster = NULL)
}
\description{
Private function
}
